feat: remove t function, clean
