--- src/routes/[[lang]]/galeria/+page.js
@@ -1,1 +1,0 @@
-export const csr = true;
