+++ src/routes/[[lang]]/galeria/+page.server.ts
@@ -1,4 +1,5 @@
-import { fetchBlogImages, type ImageMap } from '$lib/images';
+export const csr = true;
+
 import { fetchTranslation } from '$lib/api/translations';
 import type { PageServerLoad } from '../$types';
 import type { OutputData } from '../../../types/OutputData';
@@ -6,11 +7,10 @@ import type { OutputData } from '../../../types/OutputData';
 export const load: PageServerLoad<OutputData> = async ({ params }: { params: { lang?: string } }) => {
 	
 	const lang = params.lang || 'es';
-	const translations = await fetchTranslation(lang, 'contact');
-	const images = await fetchBlogImages('contact');
+	const translations = await fetchTranslation(lang, 'gallery');
 
-	return {
-		images,
+    return {
+        images: {},
 		translations
 	};
