+++ src/types/OutputData.ts
@@ -1,11 +1,13 @@
 import type { ImageMap } from "$lib/images";
 
 export type OutputData = {
-	images?: ImageMap;
+	images: ImageMap;
 	translations: any;
 };
 
-export interface OutputDataMenu extends OutputData {
+export interface OutputDataMenu {
+	translationsMenu: any;
+	translationsFooter: any;
 	lang: string;
 	defaultLocale: string;
 	locales: string[];
